+++ app/src/main/java/com/prography/android/test/hyunjung/navigation/Navigation.kt
@@ -3,17 +3,17 @@ package com.prography.android.test.hyunjung.navigation
 import androidx.annotation.DrawableRes
 import androidx.navigation.NavController
 import com.prography.android.test.hyunjung.R
-import com.prography.android.test.hyunjung.navigation.Destinations.Bookmark
 import com.prography.android.test.hyunjung.navigation.Destinations.Home
+import com.prography.android.test.hyunjung.navigation.Destinations.Random
 
 object Destinations {
     const val Home = "home"
-    const val Bookmark = "bookmark"
+    const val Random = "random"
 }
 
 class Action(navController: NavController) {
     val home: () -> Unit = { navController.navigate(Home) }
-    val bookmark: () -> Unit = { navController.navigate(Bookmark) }
+    val random: () -> Unit = { navController.navigate(Random) }
 }
 
 sealed class Screen(
@@ -21,5 +21,5 @@ sealed class Screen(
     @DrawableRes val drawableId: Int
 ) {
     object Home : Screen("home", R.drawable.ic_home)
-    object Bookmark : Screen("bookmark", R.drawable.ic_bookmark)
+    object Random : Screen("random", R.drawable.ic_random)
 }
