--- app/src/main/java/com/prography/android/test/hyunjung/ui/bookmark/navigation/BookmarkNavigation.kt
@@ -1,6 +1,0 @@
-package com.prography.android.test.hyunjung.ui.bookmark.navigation
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object BookmarkRoute
