+++ app/src/main/java/com/prography/android/test/hyunjung/ui/random/RandomScreen.kt
@@ -1,4 +1,4 @@
-package com.prography.android.test.hyunjung.ui.bookmark
+package com.prography.android.test.hyunjung.ui.random
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,11 +7,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
 @Composable
-fun BookmarkScreen(modifier: Modifier = Modifier) {
+fun RandomScreen(modifier: Modifier = Modifier) {
     Box(
         modifier = Modifier
             .fillMaxSize()
     ) {
-        Text("Bookmark")
+        Text("Random Photos")
     }
 }
