+++ app/src/main/java/com/prography/android/test/hyunjung/ui/random/navigation/RandomNavigation.kt
@@ -1,0 +1,6 @@
+package com.prography.android.test.hyunjung.ui.random.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object RandomRoute
